% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregate_br.R
\name{aggregate_br}
\alias{aggregate_br}
\title{Aggregate data in Brazil}
\usage{
aggregate_br(cov)
}
\arguments{
\item{cov}{A tibbe returned by \code{read_covid}}
}
\value{
A tibble, in which each row corresponds to a day since the first confirmed
death from COVID-19 in Brazil.
}
\description{
Takes a tibble returned by \code{read_covid} and summarises it onto
country level
}
