% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregate_br.R
\name{agregate_br}
\alias{agregate_br}
\title{Agregate data in Brazil}
\usage{
agregate_br(cov)
}
\arguments{
\item{cov}{A tibbe returned by \code{read_covid}}
}
\value{
A tibble, in which each row corresponds to a day since the first confirmed
death from COVID-19 in Brazil.
}
\description{
Takes a tibble returned by \code{read_covid} and summarises it onto
country level
}
\examples{
read_covid() \%>\% agregate_br()
}
