% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gompertz.R
\name{do_gompertz}
\alias{do_gompertz}
\title{Gompertz curve equation}
\usage{
do_gompertz(time, lambda, mu, a)
}
\arguments{
\item{time}{Integer. A vector of one or more values, counted as the number of days.}

\item{lambda}{Numeric. A single value for the lambda parameter in the gompertz equation.}

\item{mu}{Numeric. A single value for the mu parameter in the gompertz equation.}

\item{a}{Numeric. A single value for de alpha parameter in the gompertz equation.}
}
\value{
a vector of the predict number of total deaths.
}
\description{
Estimates the values of a gompertz grow process
}
\details{
This function uses the same parametrization of the gompertz curve
used in the \code{grofit} package.
}
\examples{
do_gompertz(1:100, lambda = 47, mu = 1000, a = 100000)
}
