% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_curve.R
\name{find_best_curve}
\alias{find_best_curve}
\title{Find the best gompertz curve}
\usage{
find_best_curve(grid, obs, parallel = T)
}
\arguments{
\item{grid.}{A tibble returned by \code{get_initial_params)}

\item{obs.}{A vector of observed total deaths.}

\item{parallel.}{Logical. If true, uses parallel processing from the future and
furrr packages. Defautls to true.}
}
\value{
. A named numeric vector with the values of lambda, mu and alpha from the
best curve
}
\description{
Given a grid of parameters combinations, the function estimates the gompertz
curve for each of them. Next, it calculates the sum of squared residuals and
gets the one with the lowest sum of squared residuals.
}
\examples{
\dontrun{
obs <- read_covid() \%>\% agregate_br()
obs <- obs$total_deaths

get_initial_params(lambda_min = 40, lambda_max = 80,mu_min = 800,
mu_max = 2000, a_min = 60000, a_min = 1000000) \%>\%
find_best_curve(obs)
}


}
