% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initial_params.R
\name{get_initial_params}
\alias{get_initial_params}
\title{Possible values for the Gompertz curves}
\usage{
get_initial_params(
  lambda_min,
  lambda_max,
  mu_min,
  mu_max,
  a_min,
  a_max,
  int_lambda = 1,
  int_mu = 10,
  int_a = 100
)
}
\arguments{
\item{lambda_max}{Numeric. Maximun possible value for the lambda parameter in the gompertz
equation.}

\item{mu_min}{Numeric. Minimun possible value for the mu parameter in the gompertz
equation.}

\item{mu_max}{Numeric. Maximum possible value for the mu parameter in the gompertz
equation}

\item{a_min}{Numeric. Minimun possible value for the alpha parameter in the gompertz
equation.}

\item{a_max}{Numeric. Maximum possible value for the alpha parameter in the gompertz
equation.}

\item{int_lambda}{Numeric. Increment of the sequence for the possible values of the lambda parameter.}

\item{int_mu}{Numeric. Increment of the sequence for the possible values of the mu parameter.}

\item{int_a}{Numeric. Increment of the sequence for the possible values of the alpha parameter.}

\item{lambda_min.}{Numeric. Minimun possible value for the lambda parameter in the gompertz
equation.}
}
\value{
a tibble, with every possible combination between the values in the sequences
of each parameter.
}
\description{
Given a min and and a max values for the three parameters in the gompertz equation,
generates a sequence of possible values and all the combinations of them.
}
