% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residual.R
\name{get_residual}
\alias{get_residual}
\title{Residuals for the Gompertz Curve}
\usage{
get_residual(time, obs, lambda, mu, a)
}
\arguments{
\item{time}{Integer. A vector of one or more values, counted as the number of days.}

\item{obs}{Numeric. A vector of obseverd total deaths. It must be of the same length as \code{time}}

\item{lambda}{Numeric. A single value for the lambda parameter in the gompertz equation.}

\item{mu}{Numeric. A single value for the mu parameter in the gompertz equation.}

\item{a}{Numeric. A single value for de alpha parameter in the gompertz equation.}
}
\value{
Numeric. A single value, representing the sum of squared residuals.
}
\description{
Calculate the sum of squared residuals for a gompertz curve, given
a vector of observed total deaths.
}
