% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_nls.R
\name{model_nls}
\alias{model_nls}
\title{Non linear modelling of the gompertz curve}
\usage{
model_nls(obs, best_curve)
}
\arguments{
\item{obs.}{Numeric. A vector of observed total deaths.}

\item{best_curve.}{A vector, returned from \code {find_best_curve}}
}
\value{
. A non-linear model object
}
\description{
Given a vector of observations, estimates a non linear model by
weighted least squares using the gompertz equation and the values of parameters from
the best curve found trough \code {find_best_curve} as initial values
}
