% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_nls.R
\name{predict_nls}
\alias{predict_nls}
\title{Non-linear model and prediction intervals for the best curve}
\usage{
predict_nls(m, min_time, max_time, alfa = 0.05)
}
\arguments{
\item{m}{Model, returned by \code{model_nls()}}

\item{min_time}{Integer. Starting value of the time span, counted as number of days.}

\item{max_time}{Integer. Ending value of the time span, counted as number of days.}

\item{alfa}{Numeric. Alpha level for the prediction interval}
}
\value{
A tibble.
}
\description{
Given a a non linear model by generates predictions and predictions  intervals
}
