% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_covid.R
\name{read_covid}
\alias{read_covid}
\title{Read Covid-19 Data}
\usage{
read_covid()
}
\value{
A tibble. Each row corresponds to a city/state and a date since de
beggining of the COVID-19 epidemic in Brazil.
}
\description{
Imports the most recent covid-19 data of Brazil from \url{https://brasil.io/covid19/}
}
\examples{
read_covid()
}
\author{
David Neto
}
