% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_curve.R
\name{find_best_curve}
\alias{find_best_curve}
\title{Find the best gompertz curve}
\usage{
find_best_curve(grid, obs, parallel = T)
}
\arguments{
\item{grid}{A tibble returned from \code {get_initial_params}}

\item{obs}{A vector of observed deaths}

\item{parallel}{Logical. If true, uses parallel processing for speed up.}
}
\value{
A named vector of three values: lamba, mu and a
}
\description{
Given a set of possible parameters for the gompertz curve, finds the best
candidate, which is the one with the minimal sum of least squared residuals
}
